refactor(perf): move PropertyReferenceSet to its own file for better organization
